base-password: Override whole recipe in meta-digi.
The original recipe is based in meta/recipes-core. We want to add a shadow
file for the del-image-minimal target.

The best way would be to add a bbappend overlay, however the original
recipe prepends the populate_packages function with a:

preinst = blahblah

This makes it impossible for a bbappended recipe to add anything to the
preinst script, which is needed to install the shadow file.

Hence the only solution is to duplicated the recipe and increase PR.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>