meta-digi-del: Add mesa bbappends to recipe-graphics.
The different mesa packages all provide virtual/libgl by default. However,
they all have dependencies in X.

When meta-oe is brought in, the gst-ffmpeg recipe has been appended to
depend indirectly on libgl.

On a iMX build, libgl is also provided by amd-gpu-bin-mx51, which is
set as default provider.

Still, the build shows errors because when parsing the mesa providers,
it detects that the distro does not contain the x11 feature.

This commit appends the mesa recipes only to provide libgl if x11 is
defined in the distro.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>