meta-digi-del: Add crda package.
This is based on the recipe found in the openembedded layer.

http://cgit.openembedded.org/openembedded/log/recipes/crda/crda_1.1.1.bb

It needs to compile with libnl 2.0, so we set a distro preferred version
for it.

Also, we make the recipe configurable to compile without gcrypt. The
Makefile is patched to compile without gcrypt.

Compiling with gcrypt would need the addition of the python-m2crypto-native
recipe which is not done so it is currently broken.

The compilation without gcrypt is how DEL is currently doing it so this
should be enough for our needs.

Also, the device manager is configurable and set to busybox-mdev by
default.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>