meta-digi-del: Append netbase to customize startup scripts.
Add custom /etc/init.d/networking and /etc/network/interfaces.

The networking script does the start section on the background which
inlcudes the bringing up of the interfaces. In this way DHCP does not
delay the boot up.

The interfaces file configures the wireless interface to be automatically
brought up.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>