meta-digi-arm: fix corrupted JFFS2 images in ccardxmx28js
Seems that the options Yocto was using to generate the jffs flash image
are not correct for ccardxmx28js. Specially mkfs.jffs2's padding option
'-p' was making the rootfs corrupted on second boot (as explained in
JIRA DEL-218).

Finally i decided to use the same mkfs.jffs2 parameters we were using in
DEL and those seems to be working fine.

https://jira.digi.com/browse/DEL-218 #resolve

Signed-off-by: Javier Viguera <javier.viguera@digi.com>