meta-digi-arm: Avoid udev loading atheros kernel module.
udev will attempt to load the atheros kernel modules because it defines
a modalias that matches an sdio bus event.

In a race condition with the script in if-pre-up, the module will crash.

Recent versions of udev have lost the ability to add an ignore_device
option to the rules, and mdev in del-image-minimal has no module
autoloading ability.

The commit adds a modprobe install rule that does nothing, and allows
the if-pre-up script to load the module as it used to.

Fixes https://jira.digi.com/browse/DEL-277

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>