meta-digi: Use a single Digi premirror.
This commit adds a Digi premirror to the DEL distribution file. This
premirror will be tried first. It default to build-linux.digi.conf but
will be customized for the layer public release.

This still allows for a project to be customizes with a
SOURCE_MIRROR_URL which will be tried before the premirrors.

Fixes https://jira.digi.com/browse/DEL-186

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>