Add support for variants
I have split the machine configuration in two:

-The module machine configuration include, for example ccimx53.inc
-The machine configuration (for example cimx53js.conf) which includes
the variant configuration.

This will allow for customers to create their own machine keeping the
module include file.

The variant can be configured in the local.conf file with a
MACHINE_VARIANT variable.

The variant names are made up of the interface initials as follows:

    w: wifi
    e: externel ethernel
    b: bluetooth
    1: 1-wire
    a: accelerometer
    g: GPU
    v: VPU
    128: 128MB RAM
    4k: 1024MB NAND (only for MX51 JTAG flash)

For details regarding the supported variants, each machine configuration
file contains a comment with a variants table.

https://jira.digi.com/browse/DEL-330

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>