meta-digi: move DIGI_MIRROR to meta-digi-arm
If you don't want to use meta-digi-del and use only the BSP layer
(meta-digi-arm) to build a generic image (like core-image-minimal) you
need the DIGI_MIRROR variable in this layer because some of the BSP
recipes (kernel, u-boot) use it to get the source.

So this allows to build core-image-minimal for Digi platforms without
using meta-digi-del.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>