meta-digi-del: Do not use a network manager for graphical images.
The default network manager included with the x11-sato packagegroup is
connman. It cannot coexist with ifup and /etc/network/interfaces. For
connman to work properly we would require:

- To modify netbase to leave an interfaces file with only the loopback
interface for graphical images.
- To compile the wpa_supplicant with dbus support and to leave connman the
task of starting it without a configuration file but with the dbus API (-u)
- We would need a way to run pre up scripts for all interfaces, so that the
wired interfaces can set the mac addresses passed by uboot and the wireless
interfaces can use the appropriate launch scripts.

Currently connman has no way to launch scripts before bringing an interface
up. It could be done with udev, but reportedly this does not work well
because sometimes connman brings the interface up before the udev rule is
run.

It may be that future versions of connman fix this, for example by
providing a command line tool.

For the moment it is easier to have no graphical network manager. That way
both del-image types share the network settings. In the future we may want
to revisit the need for a graphical network manager.

Fixes https://jira.digi.com/browse/DEL-387

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>