Revert "meta-digi-arm: Avoid udev loading firmware files for atheros."
This manually reverts commit ecfdf6ae2329c12857f848a854f78096a8b1810e.

Once mdev was removed as a dependency from del-image-graphical, we need
udev to load the firmware files for the wireless drivers. This job was
being done by mdev.

The reason for the commit in the first place was that both mdev and udev
were attempting to load the firmware files, so we changed the udev
rule to ignore the atheros driver. This was only needed when mdev and
udev were both running.

https://jira.digi.com/browse/DEL-450https://jira.digi.com/browse/DEL-450

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>