meta-digi-arm: Add CPX2 wireless support.
Fix the kernel wireless configuration fragment, and add a couple of kernel
patches to set up the mmc0 interface to be used with the wireless and
remove an incorrect pin configuration.

There are similar patched to this on the gateways/master branch related to
the vehicle bus adapter product.

meta-digi-arm/conf/machine/cpx2.conf

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>