Distro configuration: Allow to use the standard Yocto ADT installer.
The ADT installer compiled from "bitbake adt-installer" needs qemu images
available at adtrepo.yoctoproject.org/$VERSION. To avoid having to deploy
an adtrepo ourselves, we need to change the distro version to the upstream
version.

There are further problems with the ADT installation that need to change
the SDK_VENDOR to the poky default of pokysdk and to have the SDKPATH also
with the default value.

With this change the adt-installer is installed without errors.

https://jira.digi.com/browse/DEL-478

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>