meta-digi-del: Define xserver common virtual runtime for del-gui.
The xserver-common package in meta-openembedded replaces and conflicts
with the x11-common package in poky.

But by default poky sets the VIRTUAL-RUNTIME_xserver_common variable
in packagegroup-core-x11 to x11-common.

This causes a build error on the rootfs creation step when images with
the dev-pkgs image feature are compiled. The build error is not quite
understood, but forcing the xserver common in the distribution works
and resolves any doubt regarding to which is the package actually
being included in the target system.

xserver-common was also the package included by default in normal (non
development) builds.

https://jira.digi.com/browse/DEL-488

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>