microwindows: simplify and fix recipe
This recipe was way too complicated for what we need. Moreover there was
a build error in jenkins daily build due to this recipe:

WARNING: QA Issue: microwindows: Files/directories were installed but not shipped
  /usr/bin
  /usr/bin}
  /usr/bin}/fonts
NOTE: Tasks Summary: Attempted 2628 tasks of which 2469 didn't need to be rerun and all succeeded.

Summary: There was 1 WARNING message shown.

| Collected errors:
|  * satisfy_dependencies_for: Cannot satisfy the following dependencies for packagegroup-dey-debug-dev:
|  * 	microwindows (= 0.91-dey.r0.0) *
|  * opkg_install_cmd: Cannot install package packagegroup-dey-debug-dev.

So simplify it and fix the dependence so the jenkins job can be built.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>