meta-digi-arm: ccardimx28js: Add 3.10 kernel version.
The UBOOT_LOADADDRESS and UBOOT_ENTRYPOINT variables are only used in
the kernel recipe, so remove from the machine configuration files.

Also, use the defconfig from the kernel tree which compiles all the
support by default and expects the device tree to be configured with
the desired support. Hence, remove all the empty configuration fragments
files.

https://jira.digi.com/browse/DEL-577

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>