meta-digi: fix atheros recipes for new linux version
With the old version of the kernel (2.6.35) we build an external atheros
kernel module, while with the new kernel (3.10) the atheros driver is
included with the kernel sources.

However, the new driver needs the firmware files and the driver load
scripts that are stored with the old driver.

So put a dependence on kernel-module-atheros regardless of the kernel
version and then rework the kernel-module-atheros recipe to build the
whole driver for the old kernel or only the firmware for the new kernel.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>