kernel-module-atheros: apply load parameters to correct module
In Linux 2.6.35, the kernel module accepting the parameters on load
was 'ath6kl_sdio' while in 3.10 is 'ath6kl_core'.

So we need to apply the load parameters to either module depending on
the kernel version.

This is solved by configuring the modprobe configuration file
'/etc/modprobe.d/atheros.conf' with the parameters to the correct
module instead of passing them in the command line.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>