packagegroup-dey-audio: remove some alsa-utils dependences
Alsa-utils tools once compiled are distributed between a bunch of
different subpackages. Then 'alsa-utils' as a final package is just
a meta-package that depends on all of the different subpackages.

So instead of RDEPEND'ing on the alsa-utils metapackage, depend on just
a few of the alsa-utils subpackages.

This removes for example the 'alsa-utils-alsaconf' package that was
bringing bash into our rootfs.

Removing bash solves the problem of the lines-not-wrapped on the serial
console because now we use busybox' shell in the serial console.

https://jira.digi.com/browse/DEL-610

Signed-off-by: Javier Viguera <javier.viguera@digi.com>