meta-digi: update bluetooth to support Japan
Similar commit was done in DEL.

Japanese law limits the maximum transmit power for Bluetooth to 3 mW.
The cert field in OTP will be set to 2 when units are manufactured
for Japan.

Updated the bluez init script to configure Bluetooth for class 2
operation if the cert field is set to 2, and to delete the class 1
configuration file since it would be illegal to use it in Japan.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>