initscripts: add runtime dependency on util-linux umount
The initscripts package includes a '/etc/init.d/umountfs' script used
for reboot/poweroff purposes. That script calls the 'umount' command
but seems that the combination of busybox' umount + linux-3.x fails in
this script (it hangs without completing the reboot/poweroff).

So as a workaround use the umount command from util-linux package

https://jira.digi.com/browse/DEL-633

Signed-off-by: Javier Viguera <javier.viguera@digi.com>