packagegroup-dey-examples: add virtual/kernel dependence
The availability of some examples depends on the kernel version, so we
need to add the dependence to the packagegroup as well.

This allows to build two ccardimx28 projects one for kernel 2.x and the
other for 3.x sharing the state-cache. Otherwise the first project
builds correctly but the second one fails.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>