meta-digi-dey: removed not needed gst-plugins-good bbappend
We are now using the standard yocto kernel headers so the patch to build
the plugin with new kernel headers is not needed.

The rest of the changes in the bbappend (removing gconf dependence) were
needed in yocto-danny, but not in yocto-dylan because the gconf package
does not depend on libx11 anymore.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>