meta-digi-dey: move packagegroup definitions to distro configuration
After the upstream commit in d63c8a1597c9bd0a838cb59ca461ac432b630a90,
bitbake checks to see if the image feature has a valid packagegroup
associated.

For this check to work dey features need to be defined for all images,
not only dey images, so the definitions need to move to the distro
configuration file.

https://jira.digi.com/browse/DEL-852

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>
Signed-off-by: Javier Viguera <javier.viguera@digi.com>