meta-digi-arm: remove linux-dey-rev files
The Linux kernel revision was splitted from main bitbake recipe to its
own linux-dey-rev file so that file could be used for linux-headers
recipe to build the toolchain using custom kernel headers (from our
repository).

That is not the recommended way to build a toolchain:

http://git.yoctoproject.org/cgit.cgi/poky/commit/?id=babbf7a46acaefd9b36031483cafce053f607e66

and we are not using our kernel headers to build the toolchain anymore,
so just melt the linux-dey-rev files into the linux-dey bitbake recipes
for cleanness and easier maintenance.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>