gst-fsl-plugin: fix mfw_v4lsink on secondary display patch
Fix the patch that adds support to select the framebuffer for
mfw_v4lsink. This used to work on non-graphical images (non-X), but it
had the side effect of breaking playing accelerated videos on graphical
images (using X) with graphica applications (QT, GTK).

This fix simplifies the patch so it still adds the dual display
functionality while not breaking the graphic media player applications.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>