gst-fsl-plugin: synchronize with DEL's package
We found that the gst-fsl-plugin package from Yocto is not exactly the
same than the one we used in DEL even though they have the same
version.

We think the one we used in DEL (released by FSL via LTIB) is more
recent and has some fixes, so create a first patch that applies on top
of the one in Yocto and convert it to the one in DEL.

Also instead of appending new patches to meta-fsl-arm SRC_URI, override
completely the variable and add the patches from meta-fsl-arm layer in
our own layer.

The final result is: we use the yocto tarball, then apply a first patch
that converts it to the one in DEL/LTIB, then apply the patches from
meta-fsl-arm and finally the ones added by us.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>