gst-fsl-plugin: fix patch that sets defaults for vssconfig
Without the fix if we only configure *one* display in the vssconfig file
(i.e. we set just one entry in that file) the patch was discarding that
configuration and overwriting the configuration with dual display
defaults.

But configuring a single display in vssconfig should be doable, so fix
the patch to allow that.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>