kernel-module-atheros: workaround driver crashing on load
There seems to be a race condition in the load of the 'cfg80211' and
'ath6kl_sdio' modules. If we try to load the atheros module before the
cfg80211 is settled down, the driver crash.

The workaround is just separate the load of both modules in two separate
'modprobe' commands and add a small delay between them.

https://jira.digi.com/browse/DEL-814

Signed-off-by: Javier Viguera <javier.viguera@digi.com>