alsa-state: remove incorrect 'asound.conf' config file
The default 'asound.conf' config file is not valid for our platforms and
according to <http://www.alsa-project.org/main/index.php/Asoundrc> is
not required.

This allows to play several videos with gstreamer that were failing due
to audio issues.

It also fixes the problem of 'aplay' not playing audios with several
(more than two) channels.

https://jira.digi.com/browse/DEL-608
https://jira.digi.com/browse/DEL-872

Signed-off-by: Javier Viguera <javier.viguera@digi.com>