meta-digi-arm: set branch absolute path to overcome fetcher bug
bitbake git's fetcher uses 'ls-remote' to get the branch revision to
checkout. If the repository has two branches ending equally the 'git
ls-remote' just picks the first one regardless of that being the correct
one.

In our case we were bitten by this bug because our git repo has:

...sha1-revision...  refs/heads/android/v3.10/master
...sha1-revision...  refs/heads/v3.10/master

and the fetcher was using the android branch instead of linux's
v3.10/master.

A patch has been sent upstream but while that gets applied to Dora
maintenance branch the workaround here is to set the absolute path in
the kernel's bitbake recipe.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>