meta-digi-arm: fix u-boot version string
Building u-boot with Yocto does not show the version string in the
u-boot header. This is due to how Yocto compiles u-boot. It creates a
.scmversion file which may contain LOCALVERSION strings but in our case
it's empty. Later at build time u-boot's setlocalversion script detects
that file exists and exits with no further action (because it assumes
the file contains the version string).

https://jira.digi.com/browse/DEL-1007

Signed-off-by: Javier Viguera <javier.viguera@digi.com>