u-boot: fix build of version 2009.08 for MX5
The setlocalversion script in u-boot 2009.08 did not have the
'--save-scmversion' parameter so the build was failing.

But for that version the whole do_compile_prepend can be removed because
it's not needed at all.

While on it, update the revisions for versions 2009.08 and 2013.01 to
include a fix done in 'setlocalversion' script.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>