meta-digi-arm: add framework to set external branches for git repos
Starting with Yocto 1.6 there is a need to configure in which branch a
specific SHA1 revision is included. We need a framework to set such
branches both for internal git repos and for external ones (Github).

If the branch is not specified in the SRC_URI, the do_fetch task may
fail even though the SHA1 revision exists in the repository.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>