meta-digi: remove ALLOW_EMPTY hacks for libdigi and microwindows
Those hacks were added to fix a problem when building development images
(with 'dev-pkgs' image feature enabled). In that case the developemnt
packages (libdigi-dev and microwindows-dev) have a dependence on the
normal packages (libdigi and microwindows). But those packages are not
created by default because they are empty, thus the needed ALLOW_EMPTY hack.

Fix the problem by just resetting the runtime dependences of the -dev
packages. This allows to not create empty libdigi and microwindows
packages.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>