mplayer: remove bbappend
In Yocto Daisy mplayer is not supported. Trying to build it fails with:

ERROR: Nothing PROVIDES 'mplayer'
ERROR: mplayer2 PROVIDES mplayer but was skipped: Recipe is blacklisted: Requires newer libav which has negative D_P

This is because it has been blacklisted in commit:

http://cgit.openembedded.org/meta-openembedded/commit/?id=2e7f3ce93b2c9864037fba8cc68fde46aafac4b0

The reason is that is has a build-time dependence on libav being newer
than the one provided by Yocto.

We don't use mplayer on ccimx6 as it has accelerated video using
gstreamer, so just remove the bbappend for this release.

Probably we need to revisit this when we upgrade our ccardimx28 support
to new versions of Yocto and upstream has removed the recipe from the
blacklist.

https://jira.digi.com/browse/DEL-1075

Signed-off-by: Javier Viguera <javier.viguera@digi.com>