build.sh: purge sstate of all packagegroups between builds
Packagegroups define sets of packages to be installed in the rootfs
*depending* on the hardware variant (~MACHINE_FEATURES). For example
installing or not the bluetooth firmware depending on machine variant
having bluetooth.

It may happen that we build sequentially machines with different
features using the same sstate and thus it may happen we end with a
rootfs lacking bluetooth firmware for a bluetooth variant.

So add all packagegroups to the list of packages to purge state-cache
between platform builds.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>