mkproject.sh: prepare framework to blacklist platforms
This allows to blacklist platforms not officially supported in a DEY
release. It only works at help-level, you can actually create projects
for any platform whose config files exist, but the help of the mkproject
command will only show the officially supported platforms.

At the moment the BLACKLIST variable is not filled. This will be done in
the release branch.

https://jira.digi.com/browse/DEL-1073

Signed-off-by: Javier Viguera <javier.viguera@digi.com>