busybox: disable '-R' udhcpc option from ifupdown
udhcpc's '-R' option means 'Release IP on exit'.

This fixes a strange corner case on reboot: if you configure your network
interface to DHCP in '/etc/network/interfaces' AND you have some NFS
share mounted on your system, then running reboot command fails to
complete and it hangs on:

'Unmounting local filesystem'

This happens because on reboot TERM and KILL signals are sent to all
the processes in the system before umounting the local filesystem. When
udhcpc daemon receives the signal it exits and releases the IP (because
it was launched with '-R') leaving an unconfigured network interface
that later is unable to umount a network filesystem (NFS), resulting on
'reboot' command being unable to complete.

Removing the '-R' option from udchpc allows to kill the process without
unconfiguring the network interface.

https://jira.digi.com/browse/DEL-1125

Signed-off-by: Javier Viguera <javier.viguera@digi.com>