meta-digi-arm: set u-boot 2013.01 as preferred version
This will be the preferred version for ccardimx28. Then use platform
override to set the one for ccimx5 and ccimx6.

Not having a platform override in the machine config file allows to do
a simple version override in the project's local.conf file with:

PREFERRED_VERSION_u-boot-dey = "2009.08"

Otherwise, if we use a platform override in the machine config that
takes precedence over a normal override in the project's local.conf
and we would have needed something like:

PREFERRED_VERSION_u-boot-dey_ccardimx28 ?= "2013.01"

in the project's config.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>