meta-digi-dey: move NETWORK_MANAGER variable to x11-sato bbappend
NETWORK_MANAGER is a packagegroup-core-x11-sato recipe variable, so
override it in a bbappend instead of at a distro level.

At the same time remove a useless WEB variable from the image recipe.
WEB is a packagegroup-core-x11-sato local variable as well, and
overriding it in another recipe does not have any effect, so just
remove it.

https://jira.digi.com/browse/DEL-858

Signed-off-by: Javier Viguera <javier.viguera@digi.com>