meta-digi-dey: remove dey-gui distro
We used to have two distros:

  DEY: without X11 support
  DEY-GUI: with X11 support

After the change, we have just one DEY distro with X11 support. At the
same time add a CONFLICT_DISTRO_FEATURE to dey-image-minimal so it
cannot be built with X11 distro feature enabled. So now in order to
build 'dey-image-minimal':

DISTRO_FEATURES_remove = "x11"

needs to be added to your project's local.conf file.

https://jira.digi.com/browse/DEL-858

Signed-off-by: Javier Viguera <javier.viguera@digi.com>