meta-digi-dey: remove poky references from DEY distro
We used to have these poky references to make 'adt-installer' happy when
built using DISTRO=dey.

But this has side effects on the standalone toolchain created with
'populate_sdk'. The standalone toolchain installer has:

* default install path referring to poky instead of dey
* default toolchain version not matching our DEY version

This is misleading for customers, so rework the distro config file
prioritizing the standalone toolchain over the adt-installer.

https://jira.digi.com/browse/DEL-1003

Signed-off-by: Javier Viguera <javier.viguera@digi.com>