meta-digi-dey: blacklist adt-installer for DEY distro
The adt-installer hardcodes the distro version in URL paths to
yoctoproject website. This is ok if you control the website and the
versioning of the distro (for example using Poky as distro).

In DEY the versioning is not compatible with the content of the website,
so just blacklist the recipe. If anyone needs the adt-installer they can
download it from the Yoctoproject.

https://jira.digi.com/browse/DEL-1003

Signed-off-by: Javier Viguera <javier.viguera@digi.com>