v4l2_test: reset rotate to ROTATE_NONE on application launch
Start the application with a well known rotation value that works with
all the resolutions. Otherwise the rotation value is unknown because it
could have been changed in a previous run of the application.

Also in the kernel driver support was added to report EINVAL to
user-space in case the requested rotation is not supported. We use that
to not return from the 'v4l2_rotate' function without enabling again
the overlay preview.

https://jira.digi.com/browse/DEL-1068
https://jira.digi.com/browse/DEL-1205

Signed-off-by: Javier Viguera <javier.viguera@digi.com>
(cherry picked from commit 2f7547a1b810683b813b708ba67e53ba156a9337)