initscripts: remove runtime dependency on util-linux umount
This was added in the past (dey-1.4) as a workaround for reboot/poweroff
issues not completing properly [DEL-633]. At that time the util-linux
umount command worked while the one from busybox didn't. Now this
workaround is not needed anymore. The reboot/poweroff commands complete
correctly with busybox' umount, so just remove the runtime dependence.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>