meta-digi-arm: simplify flash images generation
We inherited from DEL the requirement to build more than one flash image
for the same rootfs contents depending on the EBS of the flash chip.

Yocto does not have support to generate more than one flash image so we
had to create functions in the image_types_digi bbclass to provide this
feature.

This commit removes that functionality and uses the standard yocto
support to generate ubifs and jffs2 flash images. The way to customize
the flash parameters is via EXTRA_IMAGECMD_jffs2 and MKUBIFS_ARGS
variables. In our case those variables are already set depending on the
different hardware variants.

https://jira.digi.com/browse/DEL-232

Signed-off-by: Javier Viguera <javier.viguera@digi.com>