meta-digi-arm: Add LDO and LDO bypass device trees.
The i.MX6 has an internal PMU with LDOs to supply voltages to the different
ARM sybdomains. There are two possible configurations, to use this internal
LDOs or to use the external PMIC to supply the voltages (LDO bypass).

The default is to boot with LDO bypass, and the device tree with the -dts
suffix will boot with the i.MX6 internal LDO.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>