atheros: don't remove wireless module on ifdown for ccardimx28
A post-down script was unloading the Atheros wireless driver
when the network interfaces were brought down.
On rare occasions the 'rmmod' command was not able to complete
and may leave the system in an unexpected state.

This patch removes the post-down script for the ccardimx28 platform
and leaves it only for the ccimx6 platform, where the unloading
is wanted to reset the driver after suspend.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-1314