meta-digi-arm: configure the console for ccardimx28js dynamically
Revert to the old way of setting the console for ccardimx28js. Configure
the console on the generic ttyS0 port name and let the device managers
create a symbolic link to the correct device at runtime.

Trying to set it statically at compile time lead to some side effects
with some packages not being recompiled after a change of the preferred
kernel version in the project's local.conf (e.g. udev-extraconf and
sysvinit-inittab).

Signed-off-by: Javier Viguera <javier.viguera@digi.com>