meta-digi-arm: fix IMAGE_FSTYPES for MX5 platforms
The IMAGE_FSTYPES variable for ccimx51js and ccimx53js was not correct,
it was taking the default one from Poky (IMAGE_FSTYPES="tar.gz")
instead of the one in 'imx-digi-base.inc' file. This happened because we
were using the weak '??=' assignement while the default in Poky was
using '?='.

Also rework a bit how this variable is set, so it is easily overriden on
project config files or initramfs recipes.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>