module-atheros: fix module not loading in kernel 2.6.35
This is a regression introduced in a74363c. The MOD_VARIANT variable was
not correctly set because in kernel 2.6.35 there is no device-tree
support.

The fix is to get the module variant from the sysfs as a fallback for
old kernels.

https://jira.digi.com/browse/DEL-1694

Signed-off-by: Javier Viguera <javier.viguera@digi.com>
(cherry picked from commit 6357d71152c38c4713fc534b78b6c2993b201aa8)