meta-digi-arm: split ath6kl firmware from kernel module
Split the ATH6KL firmware files to a different recipe and package so
they can be used regardless of what driver we are using (the external
module or the upstream driver).

Reuse the 'firmware-atheros' recipe which was creating a package with
bluetooth firmware, to create the package with wireless firmware as
well.

Also remove 'ccimx6' as compatible machine for the external atheros
module, as the module does not even compile with the kernel version we
are using (3.14).

https://jira.digi.com/browse/DEL-1858

Signed-off-by: Javier Viguera <javier.viguera@digi.com>