meta-digi-dey: Use the timestamp from the name of the image as buildname
In yocto 1.8 if the BUILDNAME is empty, it gets filled with a timestamp
during the build. However the value and format of this timestamp are
different from the ones used in the name of the image.

This change matches the formats and values, setting the value of BUILDNAME
to the timestamp used in the name of the image.

Signed-off-by: Jose Diaz de Grenu de Pedro <Jose.DiazdeGrenudePedro@digi.com>