meta-digi: remove machine specific WPA_DRIVER
Now that the MX5 platforms have been removed the only value for
WPA_DRIVER is 'nl80211' (used by Atheros wireless driver), so there is
no need to have a platform specific value.

Just initialize the value in the init-ifupdown recipe with a weak
assignment so it can be override from the project's local.conf.

https://jira.digi.com/browse/DEL-1890
https://jira.digi.com/browse/DEL-1969

Signed-off-by: Javier Viguera <javier.viguera@digi.com>