bluez-init: check on the device tree if the variant has bluetooth
The script was basing on the variant coded in the HWID to determine
if the variant had bluetooth, by comparing to an array of hard-coded variants.
This required the script to be updated with every new variant that supported
Bluetooth.
The patch checks instead if the node 'bluetooth' exists at all in the device tree
to determine if the variant supports Bluetooth.

https://jira.digi.com/browse/DEL-1933

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>