meta-digi-arm: fix busfreq include on external GPU driver
After the merge of the kernel's stability patches, the driver
uses the include for Freescale BSP beyond 3.14.28.
Despite having applied the kernel stability patches and having
a kernel version beyond 3.14.28, we still use Freescale's BSP
for 3.14.28, so let's keep the busfreq include for any 3.14.x
version.

This patch may need to be reverted if the BSP is moved to a
newer one by Freescale.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>