sysvinit: disable all cpu cores but cpu0 on halt/reboot
This is a user-space workaround for the reboot/poweroff problems we
are having in the ccimx6sbc. Seems that they are not reproducible when
only one core is enabled, so the workaround is to disable the rest of
the cores for rebooting/halting.

https://jira.digi.com/browse/DEL-2133
https://jira.digi.com/browse/DEL-2148

Signed-off-by: Javier Viguera <javier.viguera@digi.com>

(cherry picked from commit a9e36a74bb20ab38f220116ba2742dca3d80c293)