install_linux_fw_sd: workaround bad behavior of 'test' command
Old versions of u-boot 'v2013.0x' have a bug in the shell's test command
that makes structures like:

if test "${not-existing}" = "0x01"; then ...
if test -z "${not-existing}"; then ...

to fail when the checked variable does not exist.

So implement workarounds in the updater script to overcome this problem.

This is needed because the script may be used in modules with an old
pre-installed u-boot.

https://jira.digi.com/browse/DEL-2231

Signed-off-by: Javier Viguera <javier.viguera@digi.com>