dey-image-tiny: remove X11 distro feature limitation
That restriction does not make any difference in the final image
contents, because no package depending on X11 is added to the image, and
it prevents the image from being built with the default DEY
configuration (that has X11 distro feature enabled).

This also allows to build the image in the buildserver with the same job
that builds other X11 based images.

https://jira.digi.com/browse/DEL-2106

Signed-off-by: Javier Viguera <javier.viguera@digi.com>