bluez5-init: disable the HW flow control for qca6564
Add a software workaround to disable the HW flow control because the SOM
revision 3 (and older) has swapped the bluetooth TX and RX lines so we need
to operate in DTE mode which implies that the HW control lines are swapped.
This also implies reducing the baudrate to avoid the need for HW flow control.

https://jira.digi.com/browse/DEL-2611

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>