bluez5-init: add support to set qca6564 mac address
The steps to set the bluetooth MAC address is send an specific hci command
and an hci reset, so the bluetooth interface need to be up in order to
configure it.
Additionally we have generalized the way to read the MAC address from the
device tree and removed some old code for getting the MAC address in
kernel version 2.

https://jira.digi.com/browse/DUB-595

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>