meta-digi: do not use system libsdl-native
Using native libsdl from the system is showing problems in newer
versions of some operating systems (e.g. ubuntu 15.10). The default in
Yocto 2.0 has changed to build a native libsdl and use it instead of the
one in the system.

This commit synchronizes the config templates to the latest in Yocto
2.0.

https://jira.digi.com/browse/DEL-2586

Signed-off-by: Javier Viguera <javier.viguera@digi.com>